# Add MatchingFilesDatasetOp to improve performance of Dataset.list_files