--- a/tensorflow/core/api_def/base_api/api_def_MatchingFilesDataset.pbtxt
+++ b/tensorflow/core/api_def/base_api/api_def_MatchingFilesDataset.pbtxt
@@ -1,0 +1,4 @@
+op {
+  graph_op_name: "MatchingFilesDataset"
+  visibility: HIDDEN
+}