This PR aims to add `MatchingFilesDatasetOp` to improve the performance of `Dataset.list_files`, which needs to scan the entire directory tree and load the filenames into memory before the next operation starts. More details can be found at this issue (#17810).

`MatchingFilesDatasetOp` could yield the matching files in a sorted order.

The performance experiments run using [the script](https://gist.github.com/darrengarvey/ff05fbe28ab2061c101fe64353b467ff) posted by @darrengarvey , where `width` is the number of dirs at each level, and `depth` is the number of nested dirs.

Here are the initial experiment results:

when `width` = 1000 and `depth` = 2:
-- MatchingFilesDataset:
```
read first filename: time: 16.06 ms
read second filename: time: 0.52 ms
read 998 more filenames: time: 336.51 ms (0.34 ms per iteration)
```
-- Dataset.list_files():
```
read first filename: time: 253.91 ms
read second filename: time: 0.27 ms
read 998 more filenames: time: 135.77 ms (0.14 ms per iteration)
```
when `width` = 1000 and `depth` = 20:

-- MatchingFilesDataset:
```
read first filename: time: 22.16 ms
read second filename: time: 3.09 ms
read 998 more filenames: time: 2269.14 ms (2.27 ms per iteration)
```
-- Dataset.list_files():
```
read first filename: time: 1860.96 ms
read second filename: time: 0.38 ms
read 998 more filenames: time: 127.23 ms (0.13 ms per iteration)
```

The performance test-related functions (e.g., `testPerformance`) in `tensorflow/python/data/kernel_tests/matching_files_dataset_op_test.py` will be removed later.