@jsimsa 2018-09-21
Thank you for the contribution @feihugis. 

Is there a particular reason you chose to implement your logic as a dataset (as opposed to providing a more efficient implementation of the `MatchingFilesOp` kernel)? It makes more sense to me to provide an efficient MatchingFilesOp kernel. That way, your solution will support shuffling (which it currently does not) and will benefit other users of the `MatchingFilesOp` kernel.