@feihugis 2018-09-21
@jsimsa Thanks for your comments.

The issue #17810 is that `Dataset.list_files` needs to wait until the `MatchingFilesOp` scans all the files (@darrengarvey @roan100jain ). The implementation idea comes from this [comment](https://github.com/tensorflow/tensorflow/issues/17810#issuecomment-379080460) by @mrry. In my understanding, the reason is we can use the iterator over the dataset to yield the results without waiting.  For the shuffling, `ShuffleDataset` Op can be used to shuffle the dataset proposed in this solution.

I thought about how to implement this solution in `MatchingFilsOp` kernel without using `Dataset`, but did not find a way to do it. Do you have any ideas?