@feihugis 2018-09-26
@jsimsa Thanks very much for your detailed comments, which are really helpful! This PR has been revised. Could you have a look at the revisions?

For the comments about changing the existing code (e.g. variable names and iterator construction) in `list_files_dataset_op_test.py`, may I submit another PR? I also want to add a benchmark test to `list_files_dataset_op_test.py`.