@feihugis 2018-09-30
@jsimsa Thanks very much for your comments! The serialization test is added and the issues in parallelizing the call of IsDirectory are resolved. The coding style issues are resolved as well. Could you please help review these changes?