@feihugis 2018-10-01
@jsimsa Thanks for your help on this PR! If @mrry agrees to change the implementation of list_files to use the new `MatchingFilesDataset`, I would like to submit a PR for it!