@feihugis 2018-10-02
@jsimsa I'm fixing these failing tests.

For the Python 2 lint check, I run the following command on the submitted Python files. Is this the right command to reproduce the failing test (https://source.cloud.google.com/results/invocations/3b3d88eb-f0ae-4461-9840-2529e2c554b7/log)?

```
pylint --rcfile=/tmp/pylintrc tensorflow/python/data/kernel_tests/matching_files_dataset_op_test.py
```

The output is like this:
```
************* Module matching_files_dataset_op_test
C: 36, 0: Missing class docstring (missing-docstring)
C: 44, 2: Method name "_touchTempFiles" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C: 48, 2: Method name "testEmptyDirectory" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C: 57, 2: Method name "testSimpleDirectory" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C: 78, 2: Method name "testFileSuffixes" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C: 98, 2: Method name "testFileMiddles" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C:118, 2: Method name "testNestedDirectories" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C:155, 0: Missing class docstring (missing-docstring)
C:157, 2: Method name "benchmarkNestedDirectories" doesn't conform to '^(?:(?P<exempt>__[a-z0-9_]+__|next)|(?P<camel_case>_{0,2}[A-Z][a-zA-Z0-9]*)|(?P<snake_case>_{0,2}[a-z][a-z0-9_]*))$' pattern (invalid-name)
C:157, 2: Missing method docstring (missing-docstring)

------------------------------------------------------------------
Your code has been rated at 9.28/10 (previous run: 9.28/10, +0.00)
```

Besides, do you have any ideas about the errors in [this link](https://source.cloud.google.com/results/invocations/54b0d3b5-6ba5-4348-aaa6-fabbe5f6dc36/targets/%2F%2Ftensorflow%2Fpython%2Fdata%2Fkernel_tests:matching_files_dataset_op_test/log) as following?
```
======================================================================
ERROR: testEmptyDirectory (__main__.MatchingFilesDatasetTest)
Test the MatchingFiles dataset with an empty directory
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/Volumes/BuildData/tmpfs/bazel_output/_bazel_kbuilder/d1b2600cd78e76a92812a06683f5de10/execroot/org_tensorflow/bazel-out/darwin-opt/bin/tensorflow/python/data/kernel_tests/matching_files_dataset_op_test.runfiles/org_tensorflow/tensorflow/python/data/kernel_tests/matching_files_dataset_op_test.py", line 51, in testEmptyDirectory
    dataset = MatchingFilesDataset(os.path.join(self.tmp_dir, '*'))
TypeError: Can't instantiate abstract class MatchingFilesDataset with abstract methods _inputs
```