@feihugis 2018-10-02
The failed tests seem to be caused by 1) the compatibility issue of `os.makedirs` in Python 2 and Python 3; 2) the missing of `def _inputs(self)` in `MatchingFilesDataset`. These two issues have been resolved. Could @jsimsa or @qlzh727 please help run the tests again?