@jsimsa 2018-10-02
rerunning tests now