@jsimsa 2018-10-02
Ubuntu Sanity fails with:
```
FAIL: Found 1 non-whitelited pylint errors:
tensorflow/python/data/kernel_tests/matching_files_dataset_op_test.py:144: [W0622(redefined-builtin), MatchingFilesDatasetTest.testNestedDirectories] Redefining built-in 'file'
```

and the Ubuntu CC presubmit still has build failures.