@feihugis 2018-10-02
@jsimsa Sorry about the Ubuntu Sanity fails. I missed that one on my local test. Have submitted a commit to fix it.

For the [Ubuntu CC](https://source.cloud.google.com/results/invocations/1b2886fb-fc09-4abe-bfb2-adccd305ad35/targets), I did not find the fails/broken targets related to this PR. Do you have any ideas?